Updated version to particleToCurve
This version now creates polygonal spheres and parents them under the
locators. The spheres are then converted to nCloth objects and given
simple nCloth parameters to simulate bubbles.
Next tasks will be to clean up the code and insert instructions on it's
use. A quick vimeo tutorial will be created as well.