By selecting the nParticles in a scene you can grow and shrink a curve on a path.
The start of the tube follows the nParticle that existed at the time of
it's lifespan. If it's not cached it will not exist at the same time
again. The end of the curve animation is offset by 20 frames for ALL
tubes currently, this will be updated in the future to give a random
value.